requirements.txt
Django==3.2.7
django-crispy-forms==1.12.0
django-ajax-helpers==0.0.22
django-filtered-datatables==0.0.20
django-otp
qrcode
django-tab-menus==0.0.15
django-nested-modals==0.0.23
django-show-source-code==0.0.7
Sphinx
celery
redis
flower
django-celery-email
webauthn==2.0.0
django-extensions 
Werkzeug
